class ChBtnFrm moved to guiMain.py
